Flip the table on family relations
Previously, the vertices reachable by an edge were incorrectly
considered children of a vertex, and vice versa. The proper
semantics is that edges point to parents, because that's the
way key-values are inherited.

Also, the `descendants` method was misspelled.

This flips the table and turns everything upside-down.
Anyone using these methods will get broken by this change:
  * `Vertex#parents`
  * `Vertex#children`
  * `Vertex#ancestors`
  * `Vertex#decendants` [sic]