Broken specs for filtered relatives
`Vertex` includes `Comparable` and implements `<=>` with DAG ordering
semantics, which causes `==` to consider two vertices that have no
relation to be equal (with respect to ordering).

Unfortunately `Array#include?` uses `==` for equality testing, and
rspec seems to convert the `Set` actual to an `Array`, which breaks
the negative inclusion tests, and could cause false positives.

  v1 = dag.vertex
  v2 = dag.vertex

  [v1].include? v2       # => true, via v1 == v2 (via <=>)
  Set[v1].include? v2    # => false, via v1.eql? v2